# swagger.json
{
    "openapi": "3.0.0",
    "info"   : {
        "version"    : "1.0.0",
        "description": "API for Social Media Application",
        "title"      : "MINI FACEBOOK"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT" : {
                 "in"  : "header",
                 "name": "Authorization",
                 "type": "apiKey"
            }
        }
    },
    "paths": {
...
        "/api/posts/": {
            "post": {
                "tags"       : ["Posts"],
                "summary"    : "Add Posts to the site by the logged in User",
                "description": "Logged in user will now have the access to add the posts to the site! Remember the size of post should be 30Kb otherwise it throws an internal server error",
                "security"   : [
                    {"JWT" : {}}
                  ],
                  "requestBody": {
                    "required": true,
                    "content" : {                   
                    "multipart/form-data" :{
                     "schema": {
                        "type"      : "object",
                        "properties": {
                            "imageUrl" : {
                                "type": "string",
                                "format" : "binary",
                                "required": true
                            },
                            "caption" :{
                                "type" : "string",
                                "required": true
                            }
                        }
                       }
                     }
                    }
                 },
                 "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400" : {
                        "description": "Bad request"
                    },
                    "404" :{
                       "description" : "User not found"
                     },
                    "500":{
                       "description": "Internal Server Error"
                   }
                }
            },
            "get": {
                "tags"       : ["Posts"],
                "summary"    : "Get posts posted by logged in user",
                "description": "Logged in user will now have the access to check the post posted by logged in user",
                "security"   : [
                    {"JWT" : {}}
                  ],
                  "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400" : {
                        "description": "Bad request"
                    },
                    "404" :{
                       "description" : "User/ Post not found"
                     },
                    "500":{
                       "description": "Internal Server Error"
                   }
                }
            }
        },
...
        "/api/likes/":{
         "get":{
                "tags"       : ["Likes"],
                "summary"    : "Retriving the likes of the posts liked by user",
                "description": "Remember! to access this you need to be logged in via JWT token then logged in user can get the likes that he liked on posts",
                "security"   : [
                    {"JWT" : {}}
                  ],
                  "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400" : {
                        "description": "Bad request"
                    },
                    "404" :{
                       "description" : "User/ Post/ Likes not found"
                     },
                    "500":{
                       "description": "Internal Server Error"
                   }
                }
            }
        },
...
    }
}